up: Move AppSource into its own module
